Move all used code out of desktopui into sharedux. Update references accordingly.
