+++ src/AccessibilityInsights.DesktopUI/DesktopUI.csproj
@@ -84,16 +84,6 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="$(TEMP)\A11yInsightsVersionInfo.cs" />
-    <Compile Include="Enums\HighlighterMode.cs" />
-    <Compile Include="Highlighters\ImageHighlighter.cs" />
-    <Compile Include="Highlighters\OverlayHighlighter.cs" />
-    <Compile Include="Highlighters\Win32SnapshotButton.cs" />
-    <Compile Include="Highlighters\WindowHighlighterBase.cs" />
-    <Compile Include="Highlighters\GroupHighlighterItem.cs" />
-    <Compile Include="Highlighters\Highlighter.cs" />
-    <Compile Include="Highlighters\LineBorder.cs" />
-    <Compile Include="Highlighters\TextRangeHilighter.cs" />
-    <Compile Include="Highlighters\TextTip.cs" />
     <Compile Include="Properties\AssemblyInfo.cs">
       <SubType>Code</SubType>
     </Compile>
@@ -107,10 +97,10 @@
       <DependentUpon>Settings.settings</DependentUpon>
       <DesignTimeSharedInput>True</DesignTimeSharedInput>
     </Compile>
-    <Compile Include="Utility\ReferenceHolder.cs" />
     <EmbeddedResource Include="Properties\Resources.resx">
       <Generator>ResXFileCodeGenerator</Generator>
       <LastGenOutput>Resources.Designer.cs</LastGenOutput>
+      <SubType>Designer</SubType>
     </EmbeddedResource>
     <None Include="app.config" />
     <None Include="packages.config" />
@@ -144,6 +134,9 @@
     <Analyzer Include="..\packages\Microsoft.NetFramework.Analyzers.2.9.2\analyzers\dotnet\cs\Microsoft.NetFramework.CSharp.Analyzers.dll" />
     <Analyzer Include="..\packages\Text.Analyzers.2.6.4\analyzers\dotnet\cs\Text.Analyzers.dll" />
     <Analyzer Include="..\packages\Text.Analyzers.2.6.4\analyzers\dotnet\cs\Text.CSharp.Analyzers.dll" />
+  </ItemGroup>
+  <ItemGroup>
+    <Folder Include="Utility\" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Target Name="EnsureNuGetPackageBuildImports" BeforeTargets="PrepareForBuild">
