+++ src/AccessibilityInsights.SharedUx/Dialogs/TextRangeFindDialog.xaml.cs
@@ -1,17 +1,17 @@
 // Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
-using Axe.Windows.Desktop.Types;
+using AccessibilityInsights.CommonUxComponents.Dialogs;
+using AccessibilityInsights.SharedUx.Highlighting;
+using AccessibilityInsights.SharedUx.Telemetry;
 using AccessibilityInsights.SharedUx.ViewModels;
+using Axe.Windows.Core.Enums;
+using Axe.Windows.Desktop.Types;
+using Axe.Windows.Desktop.UIAutomation.Support;
 using System;
 using System.Collections.Generic;
+using System.Globalization;
 using System.Windows;
 using System.Windows.Controls;
-using Axe.Windows.Desktop.UIAutomation.Support;
-using AccessibilityInsights.DesktopUI.Highlighters;
-using Axe.Windows.Core.Enums;
-using System.Globalization;
-using AccessibilityInsights.CommonUxComponents.Dialogs;
-using AccessibilityInsights.SharedUx.Telemetry;
 
 namespace AccessibilityInsights.SharedUx.Dialogs
 {
