+++ src/AccessibilityInsights.SharedUx/Enums/HighlighterMode.cs
@@ -1,6 +1,6 @@
 // Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
-namespace AccessibilityInsights.DesktopUI.Enums
+namespace AccessibilityInsights.SharedUx.Enums
 {
     /// <summary>
     /// Highlighter Modes in Application Settings Panel
