+++ src/AccessibilityInsights.SharedUx/Highlighting/ImageOverlayDriver.cs
@@ -1,7 +1,6 @@
 // Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 using Axe.Windows.Actions;
-using AccessibilityInsights.DesktopUI.Highlighters;
 using System;
 using System.Windows.Input;
