+++ src/AccessibilityInsights/MainWindowHelpers/CCAMode.cs
@@ -2,10 +2,9 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 using AccessibilityInsights.Enums;
 using AccessibilityInsights.Misc;
+using AccessibilityInsights.SharedUx.Highlighting;
 using Axe.Windows.Actions;
 using Axe.Windows.Actions.Enums;
-using AccessibilityInsights.DesktopUI.Enums;
-using AccessibilityInsights.SharedUx.Highlighting;
 
 namespace AccessibilityInsights
 {
