+++ src/AccessibilityInsights/MainWindowHelpers/EventsMode.cs
@@ -1,17 +1,15 @@
 // Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
+using AccessibilityInsights.CommonUxComponents.Dialogs;
 using AccessibilityInsights.Enums;
 using AccessibilityInsights.Misc;
+using AccessibilityInsights.SharedUx.Enums;
+using AccessibilityInsights.SharedUx.Highlighting;
 using Axe.Windows.Actions;
 using Axe.Windows.Core.Bases;
 using Axe.Windows.Desktop.UIAutomation;
 using Axe.Windows.Desktop.UIAutomation.EventHandlers;
-using AccessibilityInsights.DesktopUI.Enums;
 using System.Collections.Generic;
-using AccessibilityInsights.SharedUx.Highlighting;
-using AccessibilityInsights.CommonUxComponents.Dialogs;
-using System.IO;
-using Newtonsoft.Json;
 
 namespace AccessibilityInsights
 {
